there was really no reason for layer to be a float

+++ include/graphics/VisualSpec.h
@@ -55,7 +55,7 @@ struct TextSpec {
 
 struct VisualSpec {
     DrawType type;
-    float layer;
+    int layer;
     union {
         PrimitiveSpec prim;
         SpriteSpec sprite;
